solve problems

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,18 +8,18 @@
 
 int main(void)
 {
-	unsigned long long int arr[98];
+	unsigned long int arr[98];
 	unsigned long int i, j1, j2, k1, k2;
 
 	arr[0] = 1;
 	arr[1] = 2;
-	printf("%llu, " ,arr[0]);
-	printf("%llu" ,arr[1]);
+	printf("%lu, " ,arr[0]);
+	printf("%lu" ,arr[1]);
 	for (i = 2; i < 92; i++)
 	{	
 		arr[i] = arr[i-1] + arr[i-2];
 		if (i != 91)
-			printf(", %llu" ,arr[i]);
+			printf(", %lu" ,arr[i]);
 	}
 	i--;
 	j1 = arr[i-1] / 1000000000;
